Remove flux estimator
