+++ Scans/Instrument.py
@@ -9,10 +9,11 @@ environment.
 """
 from __future__ import print_function
 import numpy as np
-from .Util import make_scan, make_estimator
+from .Util import make_scan
 from .Motion import Motion
 from .Defaults import Defaults
 from .Monoid import Polarisation, MonoidList
+from .Scans import estimate
 
 instrument = {"theta": 0, "two_theta": 0}
 
@@ -30,16 +31,12 @@ class MockInstrument(Defaults):
     @staticmethod
     def detector(**kwargs):
         from time import sleep
-        sleep(MockInstrument.time_estimator(**kwargs))
+        sleep(estimate(**kwargs))
         print("Taking a count at theta=%0.2f and two theta=%0.2f" %
               (instrument["theta"], instrument["two_theta"]))
         return (1+np.cos(instrument["theta"])) * \
             np.sqrt(instrument["theta"]) + instrument["two_theta"] ** 2 \
             + 0.05 * np.random.rand()
-
-    @staticmethod
-    def time_estimator(**kwargs):
-        return make_estimator(1e7)(**kwargs)
 
     @staticmethod
     def log_file():
